% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{FBIND}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{es un vector de string}

\item{b}{es otro vector de string}
}
\value{
los strings en orden alfabetico
}
\description{
FBIND
}
\examples{
fbind(a = c('hola','mundo'),b = c('foo','bar'))
}
