% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Fbind esta re bueno}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{es un vector con caracteres}

\item{b}{es un vector con otros caracteres}
}
\value{
un factor con la concatenacion de los dos vectores
}
\description{
Fbind esta re bueno
}
\examples{

fbind(a = c('hola','q','ase'), b = c('todo', 'tranka'))

}
